+++ index.js
@@ -315,13 +315,14 @@ export default class WaveFile {
    * @throws {Error} If the sample index is off range.
    */
   getSample(index) {
-    index = index * (this.dataType.bits / 8);
-    if (index + this.dataType.bits / 8 > this.data.samples.length) {
-      throw new Error('Range error');
-    }
-    return unpack(
-      this.data.samples.slice(index, index + this.dataType.bits / 8),
-      this.dataType);
+    return this.data.samples[index];
+    //index = index * (this.dataType.bits / 8);
+    //if (index + this.dataType.bits / 8 > this.data.samples.length) {
+    //  throw new Error('Range error');
+    //}
+    //return unpack(
+    //  this.data.samples.slice(index, index + this.dataType.bits / 8),
+    //  this.dataType);
   }
 
   /**
@@ -331,11 +332,12 @@ export default class WaveFile {
    * @throws {Error} If the sample index is off range.
    */
   setSample(index, sample) {
-    index = index * (this.dataType.bits / 8);
-    if (index + this.dataType.bits / 8 > this.data.samples.length) {
-      throw new Error('Range error');
-    }
-    packTo(sample, this.dataType, this.data.samples, index);
+    this.data.samples[index] = sample;
+    //index = index * (this.dataType.bits / 8);
+    //if (index + this.dataType.bits / 8 > this.data.samples.length) {
+    //  throw new Error('Range error');
+    //}
+    //packTo(sample, this.dataType, this.data.samples, index);
   }
 
   /**
@@ -347,7 +349,7 @@ export default class WaveFile {
    * @param {string} bitDepthCode The audio bit depth code.
    *    One of '4', '8', '8a', '8m', '16', '24', '32', '32f', '64'
    *    or any value between '8' and '32' (like '12').
-   * @param {!Array<number>|!Array<!Array<number>>|!ArrayBufferView} samples
+   * @param {!Array<number>|!Array<!Array<number>>|!Array<!ArrayBufferView>|!ArrayBufferView} samples
    *    The samples. Must be in the correct range according to the bit depth.
    * @param {?Object} options Optional. Used to force the container
    *    as RIFX with {'container': 'RIFX'}
@@ -362,15 +364,22 @@ export default class WaveFile {
     samples = this.interleave_(samples);
     this.updateDataType_();
     /** @type {number} */
+    
     let numBytes = this.dataType.bits / 8;
-    this.data.samples = new Uint8Array(samples.length * numBytes);
-    packArrayTo(samples, this.dataType, this.data.samples);
+    
+    //this.data.samples = new Uint8Array(samples.length * numBytes);
+    this.data.samples = new Float64Array(samples);
+    //packArrayTo(samples, this.dataType, this.data.samples);
+    
+
     this.clearHeader_();
     this.makeWavHeader(
       bitDepthCode, numChannels, sampleRate,
-      numBytes, this.data.samples.length, options);
+      //numBytes, this.data.samples.length, options);
+      numBytes, this.data.samples.length * numBytes, options);
     this.data.chunkId = 'data';
-    this.data.chunkSize = this.data.samples.length;
+    //this.data.chunkSize = this.data.samples.length;
+    this.data.chunkSize = this.data.samples.length * numBytes;
     this.validateWavHeader_();
   }
 
@@ -385,8 +394,7 @@ export default class WaveFile {
   fromBuffer(bytes, samples=true) {
     this.clearHeader_();
     this.readWavBuffer(bytes, samples);
-    this.bitDepthFromFmt_();
-    this.updateDataType_();
+    
   }
 
   /**
@@ -447,7 +455,8 @@ export default class WaveFile {
       this.fmt.numChannels,
       this.fmt.sampleRate,
       this.bitDepth,
-      unpackArray(this.data.samples, this.dataType));
+      //unpackArray(this.data.samples, this.dataType));
+      this.data.samples);
   }
 
   /**
@@ -458,7 +467,8 @@ export default class WaveFile {
       this.fmt.numChannels,
       this.fmt.sampleRate,
       this.bitDepth,
-      unpackArray(this.data.samples, this.dataType),
+      //unpackArray(this.data.samples, this.dataType),
+      this.data.samples,
       {container: 'RIFX'});
   }
 
@@ -483,17 +493,26 @@ export default class WaveFile {
     }
     this.assureUncompressed_();
     /** @type {number} */
-    let sampleCount = this.data.samples.length / (this.dataType.bits / 8);
+    //let sampleCount = this.data.samples.length / (this.dataType.bits / 8);
+    let sampleCount = this.data.samples.length;
     /** @type {!Float64Array} */
     let typedSamplesInput = new Float64Array(sampleCount);
     /** @type {!Float64Array} */
     let typedSamplesOutput = new Float64Array(sampleCount);
-    unpackArrayTo(this.data.samples, this.dataType, typedSamplesInput);
+    
+
+    //unpackArrayTo(this.data.samples, this.dataType, typedSamplesInput);
+    
+
     if (thisBitDepth == "32f" || thisBitDepth == "64") {
-      this.truncateSamples_(typedSamplesInput);
+      //this.truncateSamples_(typedSamplesInput);
+      this.truncateSamples_(this.data.samples);
     }
     bitDepthLib(
-      typedSamplesInput, thisBitDepth, toBitDepth, typedSamplesOutput);
+      //typedSamplesInput, thisBitDepth, toBitDepth, typedSamplesOutput);
+      this.data.samples, thisBitDepth, toBitDepth, typedSamplesOutput);
+    
+
     this.fromScratch(
       this.fmt.numChannels,
       this.fmt.sampleRate,
@@ -517,13 +536,15 @@ export default class WaveFile {
     } else {
       this.assure16Bit_();
       /** @type {!Int16Array} */
-      let output = new Int16Array(this.data.samples.length / 2);
-      unpackArrayTo(this.data.samples, this.dataType, output);
+      //let output = new Int16Array(this.data.samples.length / 2);
+      //let output = new Int16Array(this.data.samples.length / 2);
+      //unpackArrayTo(this.data.samples, this.dataType, output);
       this.fromScratch(
         this.fmt.numChannels,
         this.fmt.sampleRate,
         '4',
-        imaadpcm.encode(output),
+        //imaadpcm.encode(output),
+        imaadpcm.encode(this.data.samples),
         {container: this.correctContainer_()});
     }
   }
@@ -552,13 +573,14 @@ export default class WaveFile {
   toALaw() {
     this.assure16Bit_();
     /** @type {!Int16Array} */
-    let output = new Int16Array(this.data.samples.length / 2);
-    unpackArrayTo(this.data.samples, this.dataType, output);
+    //let output = new Int16Array(this.data.samples.length / 2);
+    //unpackArrayTo(this.data.samples, this.dataType, output);
     this.fromScratch(
       this.fmt.numChannels,
       this.fmt.sampleRate,
       '8a',
-      alawmulaw.alaw.encode(output),
+      //alawmulaw.alaw.encode(output),
+      alawmulaw.alaw.encode(this.data.samples),
       {container: this.correctContainer_()});
   }
 
@@ -586,13 +608,14 @@ export default class WaveFile {
   toMuLaw() {
     this.assure16Bit_();
     /** @type {!Int16Array} */
-    let output = new Int16Array(this.data.samples.length / 2);
-    unpackArrayTo(this.data.samples, this.dataType, output);
+    //let output = new Int16Array(this.data.samples.length / 2);
+    //unpackArrayTo(this.data.samples, this.dataType, output);
     this.fromScratch(
       this.fmt.numChannels,
       this.fmt.sampleRate,
       '8m',
-      alawmulaw.mulaw.encode(output),
+      //alawmulaw.mulaw.encode(output),
+      alawmulaw.mulaw.encode(this.data.samples),
       {container: this.correctContainer_()});
   }
 
@@ -1105,7 +1128,12 @@ export default class WaveFile {
   writeWavBuffer() {
     uInt16_.be = this.container === 'RIFX';
     uInt32_.be = uInt16_.be;
-    /** @type {!Array<!Array<number>>} */
+    /** @type {number} */
+    let numBytes = this.dataType.bits / 8;
+    let byteCount = this.data.samples.length * numBytes;
+    byteCount = byteCount % 2 ? byteCount + 1 : byteCount;
+    let packedSamples = new Uint8Array(byteCount);
+    packArrayTo(this.data.samples, this.dataType, packedSamples);
     let fileBody = [
       this.getJunkBytes_(),
       this.getDs64Bytes_(),
@@ -1113,8 +1141,8 @@ export default class WaveFile {
       this.getFmtBytes_(),
       this.getFactBytes_(),
       packString(this.data.chunkId),
-      pack(this.data.samples.length, uInt32_),
-      this.data.samples,
+      pack(this.data.samples.length * numBytes, uInt32_),
+      packedSamples,
       this.getCueBytes_(),
       this.getSmplBytes_(),
       this.getLISTBytes_()
@@ -1489,6 +1517,8 @@ export default class WaveFile {
     this.readDs64Chunk_(buffer);
     this.readFmtChunk_(buffer);
     this.readFactChunk_(buffer);
+    this.bitDepthFromFmt_();
+    this.updateDataType_();
     this.readBextChunk_(buffer);
     this.readCueChunk_(buffer);
     this.readSmplChunk_(buffer);
@@ -1777,7 +1807,13 @@ export default class WaveFile {
       this.data.chunkId = 'data';
       this.data.chunkSize = chunk.chunkSize;
       if (samples) {
-        this.data.samples = buffer.slice(
+        let numBytes = this.dataType.bits / 8;
+        let byteCount = chunk.chunkData.end - chunk.chunkData.start;
+        this.data.samples = new Float64Array(byteCount / numBytes);
+        unpackArrayTo(
+          buffer,
+          this.dataType,
+          this.data.samples,
           chunk.chunkData.start,
           chunk.chunkData.end);
       }